Merged PR 10143: Search block in provided vol only if not found elsewhere
Issue: Warning messages in DN logs, e.g. Exception in retrieving ReplicaInfo for block id XXXXX

Cause: The datanode searches for the requested block in all storage volumes including mounted directories. The mounted volume search writes a warning log message when the block is not found even if the block was supposed to be missing, new block for e.g.

Fix:
1. Replace a warning message with a debug log for missing blocks in mounted dirs
2. Prioritize searching the block in local dirs before looking in mounted dirs. This may avoid unnecessary lookups.