Enable tiering when FS cache is disabled
Tiered block reads fail on DNs when FS caching is disabled and the
remote store requires credentials. This happens because the DN searches
the wrong remoteFS. For e.g. there is a default provided volume created
on DN. It uses the same alias map as other provided volumes. However it
uses the default conf object. While reading a block, the volume filds
the FileRegion for the required block from alias map. However, creation
of a secured file system object will fail as the config object will not
contain the desired credentials. So the DN will fail the read.